requirements/dev.txt
aiofiles==24.1.0
alembic==1.16.1
annotated-types==0.7.0
arrow==1.3.0
astroid==3.3.10
attrs==25.3.0
aws-cdk-asset-awscli-v1==2.2.237
aws-cdk-asset-node-proxy-agent-v6==2.1.0
aws-cdk-cloud-assembly-schema==41.2.0
aws-cdk-lib==2.199.0
babel==2.17.0
backrefs==5.8
banal==1.0.6
blinker==1.9.0
boto3==1.38.26
botocore==1.38.26
build==1.2.2.post1
cacheout==0.16.0
cachetools==6.0.0
cattrs==24.1.3
cdk-nag==2.36.2
certifi==2025.4.26
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colored==2.3.0
constructs==10.4.2
coverage[toml]==7.8.2
cryptography==45.0.3
dataset==1.6.2
dill==0.4.0
events==0.5
fastcounter==1.1.0
ghp-import==2.1.0
greenlet==3.2.2
html5tagger==1.3.0
httptools==0.6.4
idna==3.10
importlib-resources==6.5.2
iniconfig==2.1.0
invoke==2.2.0
ipaddress==1.0.23
isort==6.0.1
jinja2==3.1.6
jmespath==1.0.1
jsii==1.112.0
ldappool==3.0.0
mako==1.3.10
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
multidict==6.4.4
mypy==1.16.0
mypy-extensions==1.1.0
openapi-pydantic==0.5.1
opensearch-py==2.8.0
orjson==3.10.18
packaging==25.0
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
pip-tools==7.4.1
platformdirs==4.3.8
pluggy==1.6.0
prettytable==3.16.0
prometheus-client==0.22.0
prompt-toolkit==3.0.51
psutil==7.0.0
publication==0.0.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pyfiglet==1.0.2
pygments==2.19.1
pyhocon==0.3.61
pyjwt==2.10.1
pylint==3.3.7
pymdown-extensions==10.15
pyparsing==3.2.3
pyproject-hooks==1.2.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dynamodb-lock==0.9.1
python-ldap==3.4.4
pytz==2025.2
pyyaml==6.0.2
pyyaml-env-tag==1.1
questionary==2.1.0
random-password-generator==2.2.0
requests==2.32.3
requests-aws4auth==1.3.1
requests-unixsocket==0.4.1
rich==14.0.0
s3transfer==0.13.0
sanic==25.3.0
sanic-routing==23.12.0
semver==3.0.4
sh==2.2.2
shortuuid==1.0.13
six==1.17.0
sqlalchemy==1.4.54
supervisor @ git+https://github.com/Supervisor/supervisor.git@main
tomlkit==0.13.2
tracerite==1.1.1
troposphere==4.9.2
typeguard==2.13.3
types-python-dateutil==2.9.0.20250516
typing-extensions==4.13.2
typing-inspection==0.4.1
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
uvloop==0.21.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
